+++ cardano-db/src/Cardano/Db/Schema.hs
@@ -393,6 +393,7 @@ share
     hash                ByteString          sqltype=hash32type
     txId                TxId                OnDeleteCascade
     value               Text Maybe          sqltype=jsonb
+    bytes               ByteString          sqltype=bytea
     UniqueData          hash
 
   ExtraKeyWitness
@@ -821,7 +822,8 @@ schemaDocs =
       "A table containing Plutus Data available in the blockchain, found in redeemers or witnesses"
       DatumHash # "The Hash of the Plutus Data"
       DatumTxId # "The Tx table index for the transaction where this script first became available."
-      DatumValue # "The actual data in json format"
+      DatumValue # "The actual data in JSON format (detailed schema)"
+      DatumBytes # "The actual data in CBOR format"
 
     ExtraKeyWitness --^ do
       "A table containing transaction extra key witness hashes."